+++ app/Http/Controllers/ItemController.php
@@ -115,6 +115,7 @@ class ItemController extends Controller
             $headers = new Headers();
             $headers->collection($pagination, count($items), $total)->
                 addCacheControl($cache_control->visibility(), $cache_control->ttl())->
+                addETag($collection)->
                 addSearch(Parameter\Search::xHeader())->
                 addSort(Parameter\Sort::xHeader())->
                 addParameters(Parameter\Request::xHeader())->