+++ app/Http/Controllers/ResourceController.php
@@ -97,6 +97,7 @@ class ResourceController extends Controller
             $headers = new Headers();
             $headers->collection($pagination, count($resources), $total)->
                 addCacheControl($cache_control->visibility(), $cache_control->ttl())->
+                addETag($collection)->
                 addSearch(Parameter\Search::xHeader())->
                 addSort(Parameter\Sort::xHeader());