+++ app/Http/Controllers/ResourceTypeController.php
@@ -90,6 +90,7 @@ class ResourceTypeController extends Controller
             $headers = new Headers();
             $headers->collection($pagination, count($resource_types), $total)->
                 addCacheControl($cache_control->visibility(), $cache_control->ttl())->
+                addETag($collection)->
                 addSearch(Parameter\Search::xHeader())->
                 addSort(Parameter\Sort::xHeader());