+++ app/Http/Controllers/ResourceTypeItemController.php
@@ -96,6 +96,7 @@ class ResourceTypeItemController extends Controller
             $headers = new Headers();
             $headers->collection($pagination, count($items), $total)->
                 addCacheControl($cache_control->visibility(), $cache_control->ttl())->
+                addETag($collection)->
                 addSearch(Parameter\Search::xHeader())->
                 addSort(Parameter\Sort::xHeader())->
                 addParameters(Parameter\Request::xHeader())->