+++ app/Http/Controllers/SubcategoryController.php
@@ -99,6 +99,7 @@ class SubcategoryController extends Controller
             $headers = new Headers();
             $headers->collection($pagination, count($subcategories), $total)->
                 addCacheControl($cache_control->visibility(), $cache_control->ttl())->
+                addETag($collection)->
                 addSearch(Parameter\Search::xHeader())->
                 addSort(Parameter\Sort::xHeader());