+++ resources/views/welcome.blade.php
@@ -223,6 +223,7 @@
                         <li>We have updated the transfers collection; you can filter the results by an `item` ID.</li>
                         <li>We have updated the partial transfers collection; you can filter the results by an `item` ID.</li>
                         <li>We have added additional filtering options for collections. We have added `total` and `actualised_total` range filtering for the `allocated-expense` item type and `total` range filtering for the `simple-expense` item type.</li>
+                        <li>We have added an application cache for collections; we include the ETag header however we are not yet returning a 304.</li>
                     </ul>
 
                     <h3>Changed</h3>
@@ -235,6 +236,11 @@
                         <li>We have updated the URIs for item category and subcategory assignments; the URI was singular for a collection.</li>
                         <li>We have updated the item delete endpoint; we did not return a conflict error before removing relationships.</li>
                         <li>We have tweaked the `type` for expense fields; we highlight that the supplied value should be a decimal string.</li>
+                        <li>We have moved our route validators; the route validators sit inside the `App\Request` namespace.</li>
+                        <li>We have moved our `Header` utility class; the `Header` class sits inside the `App\Response` namespace.</li>
+                        <li>We have moved our `RoutePermission` utility class; the `Permission` class sits inside the `App\Request` namespace.</li>
+                        <li>We have moved and simplified our `Parameter` classes; the `Parameter` classes sit inside the `App\Request` namespace.</li>
+                        <li>We have moved and renamed the `RequestUtility` class; the `BodyValidation` class sits in the `App\Request` namespace.</li>
                     </ul>
 
                     <h3>Fixed</h3>